+++ app_pojavlauncher/src/main/jni/Android.mk
@@ -45,6 +45,7 @@ LOCAL_SRC_FILES := \
     jre_launcher.c \
     utils.c \
     stdio_is.c \
+    java_exec_hooks.c \
     driver_helper/nsbypass.c
 
 ifeq ($(TARGET_ARCH_ABI),arm64-v8a)
